+++ R/waiter.R
@@ -14,7 +14,7 @@
 #' when the plot in \code{id} is drawn. Note the latter will only work with
 #' shiny plots, tables, htmlwidgets, etc. but will not work with arbitrary
 #' elements.
-#' @param spinners Spinners to include. By default all the CSS files for 
+#' @param spinners Deprecated argument. Spinners to include. By default all the CSS files for 
 #' all spinners are included you can customise this only that which you 
 #' need in order to reduce the amount of CSS that needs to be loaded and
 #' improve page loading speed. There are 7 spinner kits. The spinner kit
@@ -64,108 +64,37 @@
 #' @import shiny
 #' @name waiter
 #' @export
-use_waiter <- function(spinners = 1:7, include_js = TRUE){
+use_waiter <- function(spinners = NULL, include_js = TRUE){
   # to deprecate
   useWaiter(spinners, include_js)
 }
 
 #' @export 
 #' @rdname waiter
-useWaiter <- function(spinners = 1:7, include_js = TRUE){
+useWaiter <- function(spinners = NULL, include_js = TRUE){
 
   if(!isTRUE(include_js))
-    warning("include_js argument is deprecated, it is no longer needed")
-
-  # must haves
-  header <- tags$head(
-    tags$link(
-      href = "waiter-assets/waiter/waiter.css",
-      rel="stylesheet",
-      type="text/css"
+    .Deprecated(
+      "", 
+      package = "waiter", 
+      msg = "include_js argument is deprecated, it is no longer needed"
     )
+
+  if(!is.null(spinners)){
+    .Deprecated(
+      "",
+      package = "waiter",
+      msg = "spinners argument is no longer in use"
+    )
+  }
+  
+  htmltools::htmlDependency(
+    name = "waiter",
+    version = utils::packageVersion("waiter"),
+    src = "assets/waiter",
+    package = "waiter",
+    script = "waiter.js"
   )
-
-  # spinner kits
-  if(1 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/spinkit.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(2 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/css-spinners.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(3 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/devloop.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(4 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/spinners.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(5 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/spinbolt.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(6 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/loaders.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(7 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/custom.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  # add js
-  header <- shiny::tagAppendChildren(
-    header,
-    tags$script(
-      src = "waiter-assets/waiter/waiter.js"
-    )
-  )
-
-  # singleton it
-  singleton(header)
 
 }